requirements.txt
boto3==1.34.*
botocore==1.34.*
requests==2.31.*
shapely==2.0.*
python-dotenv==1.0.*
tqdm==4.66.*
